requirements.txt
torch
Pillow
transformers
requests
numpy
pandas
urllib3==1.26.16



